requirements.txt
aniso8601==9.0.1
attrs==22.1.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
Flask==2.0.2
greenlet==1.1.3
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.16.0
MarkupSafe==2.1.1
marshmallow==3.18.0
packaging==21.3
pip==21.2.4
pyparsing==3.0.9
pyrsistent==0.18.1
pytz==2022.2.1
requests==2.28.1
setuptools==58.1.0
six==1.16.0
SQLAlchemy==1.4.41
urllib3==1.26.10
Werkzeug==2.0.2
zipp==3.8.1

PyJWT~=2.5.0